all duplicate classes removed,string file now has all the classes for strings
