+++ lib/app/utils/custom_strings.dart
@@ -1,4 +1,4 @@
-class CustomStrings {
+class CustomWarningStrings {
   static const emailAlreadyRegistered =
       "OOPS !!..E mail is already registered..Try Logging in or use a different Email";
   static const registrationSuccess =
@@ -8,4 +8,13 @@ class CustomStrings {
   static const invalidCredentials = "Invalid Credentials";
   static const notRegistered = "You are not registered..";
   static const registerPrompt = "Try Registering with Us";
+}
+class SharedPrefString{
+  static const String isLoggedIn="isLoggedIn";
+} 
+class CategoryStrings{
+static const String jewelery="jewelery";
+static const String electronics="electronics";
+static const String mensClothing="men's clothing";
+static const String womensClothing="women's clothing";
 }
