--- lib/app/utils/preferences.dart
@@ -1,3 +1,0 @@
-class SharedPrefString{
-  static const String isLoggedIn="isLoggedIn";
-}
