--- lib/core/preferences_const_strings.dart.dart
@@ -1,12 +1,0 @@
-
-class SharedPrefStrings{
-static const String isLoggedIn="isLoggedIn";  
-}
-
-class CategoryStrings{
-static const String jewelery="jewelery";
-static const String electronics="electronics";
-static const String mensClothing="men's clothing";
-static const String womensClothing="women's clothing";
-}
-
