+++ lib/presentation/pages/splash_screen/splash_screen.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:task/app/utils/preferences.dart';
+import 'package:task/app/utils/custom_strings.dart';
 
 import '../../../core/routes.dart';
 import 'widget/splash_text.dart';
